+++ src/database/entity/AppraisalScore.ts
@@ -21,7 +21,7 @@ export class AppraisalScore {
   @PrimaryGeneratedColumn()
   id?: number;
 
-  @Column({ type: 'double', default: null, nullable: true })
+  @Column({ type: 'float', default: null, nullable: true })
   score?: number;
 
   @Column({
